--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "bashtest"]
-	path = bashtest
-	url = git@github.com:shawnrushefsky/bashtest.git
